fix: move findResponseHeaderValue to test util
