+++ src/lib/util.ts
@@ -247,25 +247,3 @@ export class BackoffCounter {
     this.count = 0;
   }
 }
-
-export function findResponseHeader(
-  response: TriggerResponse,
-  key: string
-): [string, string] | null {
-  const headers = response.metadata.headers || [];
-
-  const header = headers.find(tuple => {
-    return tuple[0] === key;
-  });
-
-  return header || null;
-}
-
-export function findResponseHeaderValue(
-  response: TriggerResponse,
-  key: string
-): string | null {
-  const header = findResponseHeader(response, key);
-
-  return header ? header[1] : null;
-}
